+++ public/app/features/alerting/unified/utils/access-control.ts
@@ -1,7 +1,8 @@
+import { getConfig } from 'app/core/config';
 import { contextSrv } from 'app/core/services/context_srv';
 import { AccessControlAction } from 'app/types';
 
-import { isGrafanaRulesSource } from './datasource';
+import { GRAFANA_RULES_SOURCE_NAME, isGrafanaRulesSource } from './datasource';
 
 type RulesSourceType = 'grafana' | 'external';
 
@@ -128,3 +129,12 @@ export function getRulesAccess() {
       contextSrv.hasPermission(provisioningPermissions.readSecrets),
   };
 }
+
+export const getCreateAlertInMenuAvailability = () => {
+  const { unifiedAlertingEnabled } = getConfig();
+  const hasRuleReadPermissions = contextSrv.hasPermission(getRulesPermissions(GRAFANA_RULES_SOURCE_NAME).read);
+  const hasRuleUpdatePermissions = contextSrv.hasPermission(getRulesPermissions(GRAFANA_RULES_SOURCE_NAME).update);
+  const isAlertingAvailableForRead = unifiedAlertingEnabled && hasRuleReadPermissions;
+
+  return { isAlertingAvailableForRead, hasRuleUpdatePermissions };
+};
